refactor: unify MDX + Mermaid into shared base (Closes #1477)
- Single MermaidDiagram (theme + dynamic import) for docs + architecture
- Extract large chart strings to *.charts.ts
- BASE_MDX_COMPONENTS shared by docs/terms maps
- One icon registry for docs sidebar, privacy, terms, data-rights
- No intentional behaviour change on docs/privacy/terms
