+++ src/components/architecture/PaymentFlowDiagram.tsx
@@ -2,7 +2,8 @@
 
 import { useState } from "react";
 import { BlueprintMotionSection } from "@/components/blueprint/BlueprintMotionSection";
-import MermaidDiagram from "./MermaidDiagram";
+import MermaidDiagram from "@/components/shared/MermaidDiagram";
+import { PAYMENT_SEQUENCE_CHART, PAYMENT_STATE_CHART } from "./payment-flow.charts";
 import DiagramZoomModal from "./DiagramZoomModal";
 import { Info, Maximize2 } from "lucide-react";
 
@@ -10,43 +11,9 @@ export default function PaymentFlowDiagram() {
   const [isSeqZoomOpen, setIsSeqZoomOpen] = useState(false);
   const [isStateZoomOpen, setIsStateZoomOpen] = useState(false);
 
-  const sequenceChart = `
-sequenceDiagram
-    participant Client as Client (buyer)
-    participant SWK as SWK
-    participant NestJS as NestJS (OfferHub API)
-    participant TW as TrustlessWork (escrow)
-    participant Offramp as BlindPay
+  const sequenceChart = PAYMENT_SEQUENCE_CHART;
 
-    Client->>NestJS: POST /orders (USDC reserved)
-    NestJS->>TW: deployEscrow(buyer, seller, amount)
-    TW-->>NestJS: escrowId + contractAddress
-    NestJS-->>Client: { escrowId, contractAddress }
-    
-    Client->>SWK: signTransaction(fund_escrow)
-    SWK->>TW: fund_escrow (USDC on-chain)
-    
-    Note over TW: State: ESCROW_FUNDED → IN_PROGRESS
-    
-    Client->>SWK: signTransaction(release_escrow)
-    SWK->>TW: release_escrow
-    TW-->>NestJS: webhook: escrow_released
-    NestJS->>Offramp: POST /payout (USDC → fiat)
-    Offramp-->>Client: fiat settled (SPEI / Pix / PSE...)
-  `;
-
-  const stateChart = `
-stateDiagram-v2
-    [*] --> CREATED
-    CREATED --> RESERVED
-    RESERVED --> ESCROW_FUNDED
-    ESCROW_FUNDED --> IN_PROGRESS
-    IN_PROGRESS --> COMPLETED
-    IN_PROGRESS --> DISPUTED
-    DISPUTED --> RESOLVED
-    COMPLETED --> [*]
-    RESOLVED --> [*]
-  `;
+  const stateChart = PAYMENT_STATE_CHART;
 
   return (
     <BlueprintMotionSection id="flow" className="px-6 py-24 bg-bg-base">
@@ -74,7 +41,7 @@ stateDiagram-v2
           <div className="rounded-[2.5rem] bg-bg-elevated shadow-neu-raised-l2 p-6 md:p-8 flex flex-col h-full">
             <h3 className="text-xl font-bold text-content-primary mb-6 ml-2">Transaction Flow</h3>
             <div className="relative rounded-[2rem] bg-bg-base shadow-neu-sunken p-6 flex-grow overflow-x-auto">
-              <MermaidDiagram chart={sequenceChart} className="w-full min-w-[500px]" />
+              <MermaidDiagram chart={sequenceChart} className="w-full min-w-[500px]" variant="plain" />
               <button
                 onClick={() => setIsSeqZoomOpen(true)}
                 className="absolute top-4 right-4 rounded-xl bg-bg-base shadow-neu-raised-sm p-2 hover:shadow-neu-sunken transition-all z-10"
@@ -87,14 +54,14 @@ stateDiagram-v2
               For a better view, click <Maximize2 size={11} className="inline text-theme-primary mx-0.5 -mt-0.5" /> to expand
             </p>
             <DiagramZoomModal title="Transaction Flow" isOpen={isSeqZoomOpen} onClose={() => setIsSeqZoomOpen(false)}>
-              <MermaidDiagram chart={sequenceChart} className="w-full" zoom />
+              <MermaidDiagram chart={sequenceChart} className="w-full" zoom variant="plain" />
             </DiagramZoomModal>
           </div>
 
           <div className="rounded-[2.5rem] bg-bg-elevated shadow-neu-raised-l2 p-6 md:p-8 flex flex-col h-full">
             <h3 className="text-xl font-bold text-content-primary mb-6 ml-2">Escrow State Machine</h3>
             <div className="relative rounded-[2rem] bg-bg-base shadow-neu-sunken p-6 flex-grow overflow-x-auto">
-              <MermaidDiagram chart={stateChart} className="w-full min-w-[300px]" />
+              <MermaidDiagram chart={stateChart} className="w-full min-w-[300px]" variant="plain" />
               <button
                 onClick={() => setIsStateZoomOpen(true)}
                 className="absolute top-4 right-4 rounded-xl bg-bg-base shadow-neu-raised-sm p-2 hover:shadow-neu-sunken transition-all z-10"
@@ -107,7 +74,7 @@ stateDiagram-v2
               For a better view, click <Maximize2 size={11} className="inline text-theme-primary mx-0.5 -mt-0.5" /> to expand
             </p>
             <DiagramZoomModal title="Escrow State Machine" isOpen={isStateZoomOpen} onClose={() => setIsStateZoomOpen(false)}>
-              <MermaidDiagram chart={stateChart} className="w-full" zoom />
+              <MermaidDiagram chart={stateChart} className="w-full" zoom variant="plain" />
             </DiagramZoomModal>
           </div>
         </div>
